Changed method of storing version number from keyword plugin (which was not working due to a bug), to running a command that checks the revision number of the tree and stores in a variable.
Added version of octave and loaded packages to the saved output.

Instead of only loading the needed packages, it also unloads the non-necessary packages.

Bug fix where it was possible to give negative numbers for the number of frames before start binning the data.

Bug fix where it was possible ot use less than 0 (zero) intervals when averaging the images used for find the nucleus, bleach and background.